Introduce a task to update version catalogs
This commit upgrades to Gradle 7.2 and offers a new plugin which is responsible
for updating version catalogs. It's an adhoc task which works similarly to the
old "useLatestVersions" task, but uses the information available in version
catalogs instead.

For now, not all notations are supported: only simple version notations are,
which means that if the catalog uses Gradle's "rich versions", then updating
would fail. Similarly, versions are expected to be found on the same line as
the declaration of an alias.

This will allow updating the micronaut core project to use version catalogs
instead of coordinates hardcoded in the build files, and will simplify the
generation/verification of BOM files.

It's worth noting that you must either apply the new update plugin, or the
old one, but it's NOT supported to have both since they use the same task
name: using a different task name would have broken the CI integration.

Signed-off-by: Cedric Champeau <cedric.champeau@gmail.com>