Fix Gradle 8 deprecation warnings
This commit fixes the Gradle 8 deprecation warnings by upgrading to
the latest nohttp plugin version and using the required property
for reports instead of "enabled".

In order to make future upgrades easier, it also introduces the
notion of "build environment" in the Micronaut build extension.
For now the build environment is limited to telling if it runs
within a GitHub action or not.