Avoid eager creation of javadoc/sources tasks
This commit updates the plugins to avoid the creation of tasks at configuration
time. This will make the build faster and should eventually help us getting
rid of the many afterEvaluates that bloat the build.

Unfortunately, some of those changes are potentially breaking so a compatibility
mode has been introduced, and some builds can run with the `-DstrictBuild` flag
to be stricter and avoid some configuration.