Avoid adding repositories if using strict mode
Gradle provides a way to declare repositories in `settings.gradle`, in which
case each project shouldn't declare repositories. This commit makes it possible
to use this feature, by setting the repositories mode to `FAIL_ON_PROJECT_REPOS`
in settings: if this mode is active, the convention plugin will not add any
repository automatically.