Fix nexus publishing plugin (#213)
The MicronautPublishingPlugin was trying to apply the Nexus publishing
plugin to the root project if the Nexus environment variables (or
system properties) were set. Unfortunately, this causes issues because
the evaluation of the root project is done when this happen.

Instead, the plugin will now properly be applied to the root project
if the variables are set, but it's done in a correct place (via the
settings plugin to avoid having to apply more plugins to the root
project).