Fix how the nexus repositories are added (#218)
This commit fixes yet another ordering issue. The use of `afterEvaluate` on
_any_ project to add configuration to the _root_ project is doomed to fail in
some form. Here, it was possible that the configuration of a child module was
completed before the root project was. In this case, the Nexus repository
wouldn't be added, which leads to incomplete publications!

This commit fixes the problem by making sure that the repository is added
first. It also introduces a sanity check to make sure that if the plugin is
updated, it also comes with the application of the settings plugin.