Fix javadoc/sources publishing
This commit fixes javadocs and sources publishing for modules which use
shadow jar: in this case, the publishing mechanism does NOT use standard
Gradle publications. This caused the javadocs/sources to be ignored.

In order to check that publishing works properly, there's now the
possibility to override the default sonatype URIs, in order to verify
publication locally for example.

The following env vars/system properties were added:

- allowInsecurePublishing: lets Gradle publish on a repository which
doesn't use https. Useful for dry run deployments on local repos
- SONATYPE_REPO_URI/sonatypeRepoUri : the URI of a sonatype server where
to publish releases
- SONATYPE_SNAPSHOT_REPO_URI/sonatypeSnapshotRepoUri: the URI of a nexus
repository where to publish snapshots