Fix missing configuration properties in generated docs
This commit fixes the generated documentation possibly missing
configuration properties files. To fix it, it now applies a
configuration properties plugin to Micronaut modules. This plugin
is responsible for creating tasks which will process the configuration
properties file (for each module) and generate an _outgoing configuration_
which carries those artifacts.

This makes it possible for the "main" project (the aggregating one) to
depend on those configurations in order to get the generated asciidoc
files. This effectively decouples the producer side from the consumer side,
making it less brittle and less likely to break.

It's worth noting that the same technique should be used
in other parts of the docs plugin, but this will be done later.
Basically there are many tasks which do not declare their inputs
and should be replaced with proper tasks instead. As an example,
the `zipDocs` task didn't declare any input: the consequence is
that it may, or may not contain the right docs depending on what
is called before. This is currently fixed by adding an explicit
dependency, but it should be replaced by implicit task dependencies.

Another benefit is that now all configuration properties tasks are
cacheable, which should make the build faster once we enable the
build cache.