Remove dependency on Micronaut Inject
The dependency on `micronaut-inject` causes a circular dependency
problem, which is clearly visible when using composite builds.
In practice, the dependency can be removed:

- by removing the use of `@Nullable`/`@NonNull` which do not really
bring particular value in this project
- by removing the asciidoctor property reference writer, which is
actually brought by the `micronaut-docs` project

In practice, Micronaut modules should have an annotation processor
dependency on `micronaut-docs`, which provides the writer. This
annotation processor is added automatically by the "module" plugin,
for projects which make use of it.

For others, they are typically already configured to use Micronaut
docs processor.