Rework the docs plugin (#227)
* Use the property API for the `PublishConfigurationReferenceTask`

This allows better tracking of inputs/outputs and wiring with
other tasks via implicit task dependencies.

* Express dependency between publishGuide and publishConfigurationReference

Instead of an explicit dependency between the 2, express that the page
template is actually "built by" the publishGuide task.

* Rework how docs are generated

Previously to this commit, docs generation was kind of messy. It consisted
of several tasks which copied things into a common directory, overwrote
each other's outputs, so ordering was important, leading to a number of
weird looking and hard to debug `dependsOn` and `finalizedBy` statements.

In addition, the fact of reconfiguring tasks to copy things into a common
directory was causing those tasks to be always out of date, killing performance
and making them impossible to cache.

This commit reworks the docs generation strategy to work in phases instead:

- each task constributing to docs writes to their own directory
- tasks producing intermediate docs are publishing into the `working` directory,
more specifically in a subdirectory of the `working` directory, making it
possible to see the actual output of each phase
- tasks which require the output of another task as an input are wired as
such: instead of overwriting outputs, they would not typically _copy_ the
result of the previous phase into their own working directory
- an `assembleDocs` task was added, generating the docs for a _single_ version
- this task is used as an input to the `zipDocs` task
- and eventually, an `assembleFinalDocs` task is generating the docs for
_publication on the site_, with the version dropdown selector

Last but not least, a validation task has been added, to make sure that the
generated asciidoc file do not contain unresolved directives.