Introduce Micronaut BOM plugins
This commit introduces a couple of BOM plugins, which can be used in any
Micronaut project. The first one, `io.micronaut.internal.bom` is probably
the most interesting to all projects, since it allows the creation of a
BOM for a multi-project.

Instead of having each module define how a BOM is generated, this provides
a conventional way to do it, with an extension to customize the behavior.
A typical Micronaut BOM project would only need to do this:

```gradle
plugins {
    id 'io.micronaut.build.internal.bom'
}

version projectVersion
group projectGroup
```

And this would:

- generate a BOM including all submodules, except well-known modules
(test-suite, bom, ...)
- generate a Gradle version catalog

In addition, if the module is using a Gradle version catalog, then the
BOM can include additional dependencies defined in the catalog, as long
as they are declared using the `managed.` prefix (similarly to what is
available in `micronaut-core`).

Last but not least, all generated BOMs will consistently define property
names for all modules, allowing users to override versions if needed.
The plugin also adds sanity checks (executed during the `check` phase)
to ensure that the BOM only resolves dependencies which exist in Central.

This leads us to the 2d plugin, `io.micronaut.build.internal.bom-checker`
which must be applied to a different subproject, similarly to the module
named `bom-check` in `micronaut-core`: it defines a number of tasks which
responsibility is to verify the imported BOMs. For this reason, it will
only do something if the project uses version catalogs to define the imported
BOMs (via `boms-` prefix).

This will make it less likely to write faulty BOMs like discovered in
https://github.com/micronaut-projects/micronaut-aws/pull/1216