Introduce AOT module plugin (#233)
This plugin makes it possible to build Micronaut AOT modules. A
Micronaut AOT module isn't a traditional Micronaut module: it will
not use injection, and needs to use a specific API. The plugin is
responsible for setting up the project in the correct way, and it
will automatically add the test fixtures from AOT.

This means that in a particular Micronaut repository, it is possible
to have a module which is an extension to AOT. The user would then
have to integrate the AOT module to the AOT classpath (via the Gradle
plugin for now) to benefit from the optimizations.