Merge 5.0.x branch into master (#235)
* Remove dependency on Micronaut Inject (#226)

The dependency on `micronaut-inject` causes a circular dependency
problem, which is clearly visible when using composite builds.
In practice, the dependency can be removed:

- by removing the use of `@Nullable`/`@NonNull` which do not really
bring particular value in this project
- by removing the asciidoctor property reference writer, which is
actually brought by the `micronaut-docs` project

In practice, Micronaut modules should have an annotation processor
dependency on `micronaut-docs`, which provides the writer. This
annotation processor is added automatically by the "module" plugin,
for projects which make use of it.

For others, they are typically already configured to use Micronaut
docs processor.

* Rework the docs plugin (#227)

* Use the property API for the `PublishConfigurationReferenceTask`

This allows better tracking of inputs/outputs and wiring with
other tasks via implicit task dependencies.

* Express dependency between publishGuide and publishConfigurationReference

Instead of an explicit dependency between the 2, express that the page
template is actually "built by" the publishGuide task.

* Rework how docs are generated

Previously to this commit, docs generation was kind of messy. It consisted
of several tasks which copied things into a common directory, overwrote
each other's outputs, so ordering was important, leading to a number of
weird looking and hard to debug `dependsOn` and `finalizedBy` statements.

In addition, the fact of reconfiguring tasks to copy things into a common
directory was causing those tasks to be always out of date, killing performance
and making them impossible to cache.

This commit reworks the docs generation strategy to work in phases instead:

- each task constributing to docs writes to their own directory
- tasks producing intermediate docs are publishing into the `working` directory,
more specifically in a subdirectory of the `working` directory, making it
possible to see the actual output of each phase
- tasks which require the output of another task as an input are wired as
such: instead of overwriting outputs, they would not typically _copy_ the
result of the previous phase into their own working directory
- an `assembleDocs` task was added, generating the docs for a _single_ version
- this task is used as an input to the `zipDocs` task
- and eventually, an `assembleFinalDocs` task is generating the docs for
_publication on the site_, with the version dropdown selector

Last but not least, a validation task has been added, to make sure that the
generated asciidoc file do not contain unresolved directives.

* Introduce a BOM plugin (#228)

* Move the extension creation to its own plugin

This removes an unecessary dependency between the dependency
updates plugin and the base plugin.

* Introduce Micronaut BOM plugins

This commit introduces a couple of BOM plugins, which can be used in any
Micronaut project. The first one, `io.micronaut.internal.bom` is probably
the most interesting to all projects, since it allows the creation of a
BOM for a multi-project.

Instead of having each module define how a BOM is generated, this provides
a conventional way to do it, with an extension to customize the behavior.
A typical Micronaut BOM project would only need to do this:

```gradle
plugins {
    id 'io.micronaut.build.internal.bom'
}

version projectVersion
group projectGroup
```

And this would:

- generate a BOM including all submodules, except well-known modules
(test-suite, bom, ...)
- generate a Gradle version catalog

In addition, if the module is using a Gradle version catalog, then the
BOM can include additional dependencies defined in the catalog, as long
as they are declared using the `managed.` prefix (similarly to what is
available in `micronaut-core`).

Last but not least, all generated BOMs will consistently define property
names for all modules, allowing users to override versions if needed.
The plugin also adds sanity checks (executed during the `check` phase)
to ensure that the BOM only resolves dependencies which exist in Central.

This leads us to the 2d plugin, `io.micronaut.build.internal.bom-checker`
which must be applied to a different subproject, similarly to the module
named `bom-check` in `micronaut-core`: it defines a number of tasks which
responsibility is to verify the imported BOMs. For this reason, it will
only do something if the project uses version catalogs to define the imported
BOMs (via `boms-` prefix).

This will make it less likely to write faulty BOMs like discovered in
https://github.com/micronaut-projects/micronaut-aws/pull/1216

* Make sure that the BOM checker plugin applied the base plugin

This makes sure that the `clean` and `check` tasks are available.
This fixes an issue with build caching, where potentially the output
of BOM file verification wasn't cleaned.

* Prepare branch 5.0.x

* Add test retry plugin (#232)

This will let Gradle Enterprise detect flaky tests.

* Introduce AOT module plugin (#233)

This plugin makes it possible to build Micronaut AOT modules. A
Micronaut AOT module isn't a traditional Micronaut module: it will
not use injection, and needs to use a specific API. The plugin is
responsible for setting up the project in the correct way, and it
will automatically add the test fixtures from AOT.

This means that in a particular Micronaut repository, it is possible
to have a module which is an extension to AOT. The user would then
have to integrate the AOT module to the AOT classpath (via the Gradle
plugin for now) to benefit from the optimizations.

* Use lazy configuration for the dependency updates tasks