Fix publication of snapshots
Publication of snapshots was driven by the root project version.
With the changes in 5.0.0, it's not always the case that the
root project had a version.

This commit fixes the problem by making sure that the project
version is set on all subprojects, including the root. This
makes use of the `projectVersion` property which is expected
to be found in `gradle.properties`. A sanity check was added
to make sure that this property is set when publishing.

In addition, there is now the ability to set a project group,
using the `projectGroup` property, which should make it
redundant to define a convention plugin which declares the
group and version.