Fix how aggregated javadocs are produced (#241)
* Fix how aggregated javadocs are produced

Before this commit, Javadoc aggregation was generated by accessing other
project's state directly. While being much simpler to write, this was
causing deprecation warnings in the build, regarding configurations
being resolved out of context.

The fix wasn't trivial to implement, because we need a generic fix
which supports includes and excludes. Therefore, the way the aggregation
works is now by applying a couple of plugins:

- on a module which needs to be aggregated, we apply a plugin which is
responsible for setting up a couple of outgoing configurations: one
for the _sources_ which need to participate in javadocs, and another
one with the _classpath_. This is `AggregatedJavadocParticipantPlugin`
- on the aggregating project, we also need a couple of configurations
which are this time resolved, for the sources and classpath, but we
also need a disambiguation rule for the aggregated classpath, so that
there's no ambiguity between the compile classpath and the javadoc
"compile" classpath.

This also means that this process is also slower, since we need
to produce an artifact on the aggregated projects which is the sources
with the include/exclude filters applied, instead of getting a "view"
of the existing sources, which would be sufficient. Currently this
is implemented as a "sync" copy (not a zip).

* Add sanity check for javadocs linking to Micronaut projects URI