Add ability to publish scans on demand
This commit adds a Gradle property which can be set to disable automatic
publishing of build scans. If this property is set to `true`, then
publishing a build scan requires to add `--scan` to the CLI invocation.

There are 2 options to pass this option:

1. via the CLI, by adding `-PpublishScanOnDemand=true`
2. by adding `publishScanOnDemand=true` to your local Gradle properties
(the file at `~/.gradle/gradle.properties`).