Allow version catalogs to include aliases from nested catalogs (#281)
* Allow version catalogs to include aliases from nested catalogs

Our BOM plugin lets us declare which BOMs should be imported, via
the `boms.` property prefix in the `libs.versions.toml` file. For
example:

```toml
boms-micronaut-gcp = {
    module = "io.micronaut.gcp:micronaut-gcp-bom",
    version.ref = "managed-micronaut-gcp"
}
```

would trigger the creation of a BOM which dependency
management section _imports_ the gcp BOM. However, if
we move from entries which were previously declared
in the catalog, and are now defined in the imported
BOM, we have a problem. For example, if the catalog
used to contain this:

```toml
managed-micronaut-gcp = {
    module = "io.micronaut.gcp:micronaut-gcp-function",
    version.ref = "managed-micronaut-gcp"
}
```

and that we remove the entry because we're moving
to importing the BOM via `boms.`, then it's a breaking
change for users, who wouldn't be able to use the notation:

```gradle
dependencies {
   implementation(mn.micronaut.gcp.function)
```

in their build files, because that entry would be removed.

This commit fixes the problem by introducing an _inlining_
mechanism. Basically, the idea is that when we import a BOM,
we try to see if it's published with a version catalog
(which is the case for the Micronaut BOMs). If so, then we
_inline_ the entries of the imported catalog into the current
one, without declaring a version.

The net result, in the example above, is that the alias is
effectively still present in the Micronaut BOM, so users
can continue to use the alias as before. Of course, this only
works if, and only if, the aliases of the nested catalog are
actually the same as the removed ones. From my tests, this
looks to be often the case. A noticeable difference is that
for `mn.micronaut.gcp` you'd have to use `mn.micronaut.gcp.common`
instead (because of that alias difference). This is still
a breaking change, but a reasonable tradeoff.

Currently, *only* aliases which start with `micronaut` are
inlined. This also means that if the imported BOM is actually
an external project BOM, like, say, the jackson BOM, then we
don't do anything and don't add entries to the catalog
automatically. Even if the external BOM is using a catalog,
we won't automatically import entries, unless, by accident,
they start with `micronaut`.

See https://github.com/micronaut-projects/micronaut-core/pull/6987

* Lift restriction on aliases to inline

Instead of filtering by entries starting with `micronaut`, there's
now a property which lets the user declare which aliases shouldn't
be imported.

Added functional test.

* Add warning in case an alias is already declared