Improve how we decide to include a project in BOM
Before this change, the only way to exclude a project from the BOM
was to use the `excludeXXX` methods on `micronautBom` extension.
With this change, by default, we will now only include projects
which apply the `publishing` plugin, that is to say the projects
which are supposed to be published on Maven Central.

In case this causes problems, it is possible to disable inference
by setting the appropriate property:

```gradle
micronautBom {
    inferProjectsToInclude = false
}
```

The plugin will now also show which projects are included
when creating the POM file.