Improve version property lookup
This commit improves how Micronaut versions are read, so that we can
consistently use version catalogs. Previously to this change, some
versions still had to be declared in `gradle.properties`, which was
inconsistent with the use of the version catalog for other dependencies.

With this change, it is possible to declare everything in the catalog,
including the special versions `micronaut`, `micronaut-docs`, `micronaut-test`,
`groovy` and `spock`.

This change in made in a compatible way: if the version is defined in the catalog
it will be fetched from it, otherwise it will get the old version from the
gradle properties file.