Fix spotless dependency (#346)
I am not quite sure why we have a dependency between checkstyle and
spotless. However since it's there, let's fix it so that it works
properly in case the spotless plugin isn't applied.

This commit fixes the case where the code quality plugin is applied
on a project which doesn't apply the spotless plugin, which might
be the case for "test projects" for example.