Fix binary compatibility checks
This commit fixes a couple of issues with the binary compatibility
checks:

- first, the "baseline" task was cacheable, but didn't take into
account the actual contents of the JSON file returned by GH releases.
This means that if a release was made in between, we could compare
against an older baseline than expected.
- second, this commit introduces a rule to ignore Java bridge methods
and synthetic classes

The reason for the 2d change is that we moved from releasing with Java 8,
to releasing with Java 11 and targetting Java 8 bytecode. However the
Java 11 compiler may generate different synthetic code than the Java 8
one. So we simply ignore those internal implementation details.