Run findBaseline only once
When we build core, we currently make 20 requests to the GitHub api to find the baseline.

This gives us a top limit of 50 builds per hour (if my calculations are correct) as authenticated users
have a rate limit of 1000 requests per hour.

This change moves the findBaseline task up to the root project so that it is only called once
and creates baseline.txt in the root build folder.

We then use this single file as an input to the japiCmp tasks in each module project.

I'm not a massive fan of how I find if the task currently exists in the root project, but I
could not find another way to get a TaskProvider for it other than calling named() and catching
the exception

This builds on top of the authentication PR #366