Cache find baseline for an hour
The previous implementation required fetching the remote list of
releases before determining if the task was cacheable or not. This
caused rate limiting issues on GitHub (while this was in theory
the right thing to do).

To workaround limits, we're replacing the JSON input with a
timestamp corresponding to the current hour, so we'd fetch at most
once per hour.