Check version catalog compatibility
This plugin adds a "binary" compatibibility check for the published
version catalogs. It will check that the newly published catalogs
doesn't miss entries which existed previously. If the change is
intentional, it is possible to explicitly accept regressions by
adding them to the `suppressions` block of the `micronautBom`
extension.

The tests are using "real" catalog files, the baseline is 3.5.3
and the baseline is the current 3.6.0 branch state. It already
shows that we discover problems using this.