Use providers instead of relying on `afterEvaluate`
This commit updates how default dependencies are added, so that they
use providers instead of using `afterEvaluate`. Versions which were
hardcoded, for example for `cglib`, are now also configurable.

In addition, this commit introduces a new configuration, `globalBoms`,
which is used to declare the BOMs which have to be applied "globally",
in the sense that they must be applied to both the annotation processor
path and the various compile and runtime classpaths.

This makes it easier for additional configurations to opt-in to use
of BOMs by simply extending from this configuration, instead of having
to duplicate the BOM declarations in every configuration.