Default to Java 17
This commit makes the build plugin configure Java 17 by default.
It deprecates the use of `sourceCompatibility` and `targetCompatibility`
on the `micronautBuild` extension, in favor of using Gradle's toolchain
support instead.

If a project is not ready for Java 17, which is likely to happen for
current Micronaut modules, it is possible to override the default
Java version by setting the `micronautBuild.javaVersion` property.

If a project is _today_ using the `micronautBuild.sourceCompatibility`
property (or `targetCompatibility`), then toolchain support isn't used,
and a deprecation warning is issued during the build.