Include a pom.xml file into the published jars (#416)
This commit adds the inclusion of the generated `pom.xml` file
into the `META-INF/maven/group.id/artifact.id` directory of the
published jars. While this file can be fetched externally, some
vulnerability scanners actually look for this file.

It doesn't guarantee that users would use the exact same version
of dependencies, but it is consistent with what Maven would do,
for the better or worse.