Support SNAPSHOT versions in BOM check
Previously, we were not handling snapshots correctly: on remote repositories,
the POM file ending with `-SNAPSHOT.pom` may not exist. This fixes the download
by looking at the `maven-metadata.xml` file.

In order for us to handle snapshots correctly, we also must re-download the
snapshots every time (we don't perform caching for now).