Skip some checks if Micronaut BOM (#432)
If the current POM file being checked is a Micronaut BOM, then we
assume that we don't want to check that it should only "talk" about
dependencies in the same group. Otherwise, we have many validation
failures because basically all of our BOMs do this.

Only BOMs which are from other sources are verified, and need an
explicit suppression of error.