Use the new core BOM instead of the platform BOM (#437)
* Bump version

* Change the target BOM coordinates

Before this change, if a project used `importMicronautCatalog`,
then the BOM which would be imported would be the Micronaut BOM,
that is to say `io.micronaut:micronaut-bom`. With this change,
if the target Micronaut version is 4+, then we will import the
core catalog instead.

It is possible for modules to import other catalogs (BOMs)
in their own namespace. For example, imagine a project which
wants to use the `core` dependencies as well as those from
the `aws` project, then in the settings file you can write:

```gradle
micronautBuild {
    // imports the "core" catalog as "mn"
    importMicronautCatalog()
    // imports the "aws" catalog as "mnAws"
    importMicronautCatalogWithGAV("io.micronaut.aws:micronaut-aws-bom:3.9.3")

    // alternatively, if the project uses a version catalog and declares
    // the BOM using alias `boms-micronaut-aws`, then you can do this:
    importMicronautCatalog("boms-micronaut-aws")
}
```

* Fix BOM artifact id depending on micronaut version