Add support for "conventional" project names (#477)
This commit adds the ability to use conventional project names,
that is to say project names which start with `micronaut-`. This
allows having publications which artifact id match the project name.
Not having this is source of trouble for using composite builds,
since Gradle wouldn't be able to determine the substitutions
automatically. Changing the project names to match the artifact
id fixes this issue, at the cost of some drawbacks:

   - full task paths will now use `:micronaut-xxx` as a prefix,
instead of just `:xxx`
   - project dependencies will be broken when this feature is
enabled. For example, a dependency on `project(":foo")` will
have to be rewritten to `project(":micronaut-foo")`

The change is _not_ applied to the root project (which is not
published in any case), which will let us preserve the Gradle
Enterprise history for that project. It is not applied either
to projects which name starts with `test-suite`.

This feature is _disabled by default_ for now, it needs to
be enabled in the `settings.gradle` file, under the build
extension:

```gradle
micronautBuild {
    useStandardizedProjectNames = true
}
```