Fix race condition when verifying BOM (#486)
When verifying that a BOM file only refers to dependencies which
exist on remote repositories, we perform parallel verification of
all artifacts in the dependency graph, in order to make verification
faster.

However, it is possible that the same dependency appears multiple
times in the graph, in which case we could try to download the
same file concurrently, which is fine, but we would also try to
copy the result into our local cache. This step could fail because
another thread could be writing the same file.

In order to protect from this, we now make sure that concurrent
requests for the same artifact are blocked.