Use Maven Central for fetching release list
GitHub is giving us troubles with API rate limits. So instead of
using it to fetch the release list, this commit makes use of Maven
Central. In order to avoid spamming Maven Central, we also make
sure to perform a single request per host, and we also cache the
result for the lifetime of the build.

This also bumps the Gradle Enterprise plugin version so that we
can benefit from test distribution if we release 5.4.x.