Make it possible to run tests on specific JDK
This commit adds the ability to run tests on a specific JDK, by
making the `java.vendor` an input to the test tasks. In order to
reduce the impact on caching for other users, this is opt-in: it
will require setting the `MICRONAUT_TEST_USE_VENDOR` environment
variable to `true` in a workflow, in which case the tests will
have the vendor as an additional input.

Important: this will only work for modules which apply the
`io.micronaut.build.internal.common` plugin (which is the
case for `io.micronaut.build.internal.base-module` and
`io.micronaut.build.internal.module`), so it is likely
that projects which also include test suites or benchmarks
will have to do something similar if they want tests to be
executed on a different JDK.

Fixes #494