Fix binary compatibility checks
When standardized project names are in used, binary compatibility checks
fail because the `micronaut-` prefix is added to the artifact id, resulting
in a double `micronaut-micronaut-` prefix. This commit fixes that, but by
doing so, it also introduces another bug which is related to Gradle preventing
the resolution of a library which corresponds to the current project, in a
different version.

In order to workaround this bug, we use the root project detached configurations,
which isn't great, because it reaches out to another project's state.