Introduce pre-release check task
The `preReleaseCheck` task is designed to be executed when a module
is still in SNAPSHOT state, but we want to make sure that it is ready
for release.

For now, the task will only trigger the verification of the generated
POM file (and their transitive dependencies), in order to check that
a release doesn't contain a dependency on a SNAPSHOT.

It is possible to bypass the failures by running with the
`micronaut.fail.on.snapshots` system property set to `true`.