Convert the publishing plugin to Java
This commit is a potential breaking change: it converts the publishing
plugin to Java (to make things easier to maintain), and removed in
the process support for the `extraPomInfo`, `startPomInfo` and
`pomInfo` closures which were used to tweak the generated pom files. If
tweaking is required, then this must be done via the type-safe API
instead.

Relates to #536