Make BOM generation compatible with the configuration cache
This was a bit tricky, because we defer configuraton of the model
until the task is actually required (via our model builder). This
caused a number of things to be captured by the configuration cache,
like the extension, version catalog, or even project.

The solution was to use providers, so that the result of the provider
can be serialized.