Prefer project modules in conflict resolution (#542)
We have seen releases failing when generating the javadocs. The reason
is that there's conflict resolution happening between versions like
`-M1` and `-SNAPSHOT`, where `-SNAPSHOT` wins. The consequence is
that when we compute the javadoc classpath, we don't resolve project
dependencies, but binaries, which miss the javadoc classpath variant.

In general, we should always prefer the project mdoules during
conflict resolution: it is extremely unlikely we intentionally
want to use a binary version of a project instead of the local
version. Should this happen, the project can override the configuration
by calling:

```
configurations.all {
   resolutionStrategy {
      failOnVersionConflict()
   }
}
```
then explicitly handle conflict resolution. This
is arguably a bit painful, but there's no method
on `ResolutionStrategy` to revert to the default
behavior.