Fix compatibility with Gradle's configuration cache (#538)
* Convert the publishing plugin to Java

This commit is a potential breaking change: it converts the publishing
plugin to Java (to make things easier to maintain), and removed in
the process support for the `extraPomInfo`, `startPomInfo` and
`pomInfo` closures which were used to tweak the generated pom files. If
tweaking is required, then this must be done via the type-safe API
instead.

Relates to #536

* WIP: Make catalog publishing compatible with config cache

* Make BOM generation compatible with the configuration cache

This was a bit tricky, because we defer configuraton of the model
until the task is actually required (via our model builder). This
caused a number of things to be captured by the configuration cache,
like the extension, version catalog, or even project.

The solution was to use providers, so that the result of the provider
can be serialized.

* Fix compatibility of the docs generation with config cache

* Fix compatibility of the createReleasesDropdown task with CC

It seems that the Groovy closures which are serialized by the
configuration cache do not resolve properly the delegates, leading
to strange errors like `logger` not found or `versionsJson` not
found.